test: add unit tests for actor use cases
